Update delete logic

+++ backend/controllers/product.controller.js
@@ -48,6 +48,11 @@ export const updatedProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -55,6 +60,6 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log("Error in deleting product", error.message);
-    res.status(500).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server error" });
   }
 };
